// server/src/plugins/organizations/gql/generated.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  DateTime: { input: any; output: any; }
  JSON: { input: any; output: any; }
  Money: { input: number; output: number; }
  Point: { input: any; output: any; }
  Upload: { input: any; output: any; }
};
// ...
export type CreateOrganizationAddressInput = {
  city: Scalars['String']['input'];
  country: Scalars['ID']['input'];
  fullName: Scalars['String']['input'];
  location?: InputMaybe<PointInput>;
  organization: Scalars['ID']['input'];
  phoneNumber: Scalars['String']['input'];
  postalCode: Scalars['String']['input'];
  province?: InputMaybe<Scalars['ID']['input']>;
  streetLine1: Scalars['String']['input'];
  streetLine2: Scalars['String']['input'];
};
// ...
export type PointInput = {
  latitude: Scalars['Float']['input'];
  longitude: Scalars['Float']['input'];
};
// ...
export type UpdateOrganizationAddressInput = {
  city?: InputMaybe<Scalars['String']['input']>;
  country?: InputMaybe<Scalars['ID']['input']>;
  fullName?: InputMaybe<Scalars['String']['input']>;
  id: Scalars['ID']['input'];
  location?: InputMaybe<PointInput>;
  organization?: InputMaybe<Scalars['ID']['input']>;
  phoneNumber?: InputMaybe<Scalars['String']['input']>;
  postalCode?: InputMaybe<Scalars['String']['input']>;
  province?: InputMaybe<Scalars['ID']['input']>;
  streetLine1?: InputMaybe<Scalars['String']['input']>;
  streetLine2?: InputMaybe<Scalars['String']['input']>;
};